requirements.txt
mock
SOAPpy
PyYAML
